Refactor to separate Qt stuff
